Add files via upload
A web based platform that provides rank based access control to military resources of the Indian Armed Forces. Involves dynamic creation of user accounts and storing them all in a SQL database incorporated with PHP. It provides a centralised view of the resources and its officials to the admin of this system.